v 2.5/requirements.txt
streamlit~=1.33.0
requests~=2.31.0
Pillow~=10.3.0
numpy~=1.24.3
tensorflow~=2.16.1
plotly~=5.20.0
openai~=1.16.2
keras~=3.1.1
python-dotenv~=1.0.1
dotenv
streamlit_lottie
Werkzeug==3.0.2
absl-py==2.1.0
aiofile==3.8.8
aiofiles==23.2.1
aiohttp==3.9.3
aiopath==0.7.7
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
caio==0.9.13
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
craiyon.py==1.2.0
decorator==5.1.1
distro==1.9.0
executing==2.0.1
flatbuffers==24.3.25
frozenlist==1.4.1
gast==0.5.4
gitdb==4.0.11
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib-metadata==7.1.0
ipython==8.23.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.1.1
libclang==18.1.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.4.0
multidict==6.0.5
namex==0.0.7
numpy==1.26.4
oauthlib==3.2.2
openai==1.16.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
parso==0.8.4
pip==24.0
plotly==5.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==5.26.1
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.6.4
pydantic-core==2.18.0
pydeck==0.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
setuptools==69.2.0
six==1.16.0
smmap==6.0.0
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.33.0
streamlit-lottie==0.0.5
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing-extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.28.0
watchdog==4.0.0
wcwidth==0.2.13
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
